+++ src/actions/auth.js
@@ -1,0 +1,27 @@
+import { push } from 'react-router-redux';
+
+import * as action from './types';
+
+// const ROOT_URL = '/api';
+// const LOGIN_URL = `${ROOT_URL}/signin`;
+// const SIGNUP_URL = `${ROOT_URL}/signup`;
+
+export const signInUser = () => {
+    return dispatch => {
+        dispatch({
+            type: action.USER_LOGIN_SUCCESS
+        });
+
+        dispatch(push('/cards'));
+    }
+};
+
+export const signOutUser = () => {
+    return dispatch => {
+        dispatch({
+            type: action.USER_LOGOUT
+        });
+
+        dispatch(push('/'));
+    }
+};
