+++ src/actions/types.js
@@ -11,8 +11,8 @@ export const FETCH_CARD_FAILED = 'FETCH_CARD_FAILED';
 export const FETCH_TRANS_FAILED = 'FETCH_TRANS_FAILED';
 export const FETCH_TRANS_SUCCESS = 'FETCH_TRANS_SUCCESS';
 
-export const CHANGE_ACTIVE_CARD = 'CHANGE_ACTIVE_CARD';
+export const ACTIVE_CARD_CHANGE = 'ACTIVE_CARD_CHANGE';
 
 export const MOBILE_PAY_SUCCESS = 'MOBILE_PAY_SUCCESS';
 export const MOBILE_PAY_FAILED = 'MOBILE_PAY_FAILED';
-export const MOBILE_PAY_WISH_TO_REPEAT = 'MOBILE_PAY_WISH_TO_REPEAT';
+export const MOBILE_PAY_REPEAT = 'MOBILE_PAY_REPEAT';
